refactor: split heading logic
